First version of message popup working
For now it is possible to set the minimum and maximum time in between changes and a list of messages. After a random amount of time it will choose a random message and change the UI text into it.

Next is figuring out how to make the textbox appear and disappear again.